add the offline tool to add lon_lat_ps to IODA and updated each jedi yaml to use that as category variable (#259)
1. A new variable is added to IODA in an offline tool which is provided
in this PR. The offline tool is only a temporary solution and the best
place to put this will likely be in a python converter. This tool is
based on the offline domain check.
2. The new variable `MetaData/longitude_latitude_pressure` is now used
as the `category variable` to check for duplicate obs matching the
functionality that is in GSI using `l_closeobs` option.