Add `ObsError` group to radiance IODA obs via the offline domain check (#287)
This PR addresses #274. To prevent the empty `DerivedObsError` group
from being created and used during LETKF, we need to include `ObsError`
in the input IODA data for radiance obs. The bufr2ioda converters for
radiance data are still in active development and are not ready to be
committed to RDASApp. Thus, the easiest solution for us is to add
`ObsError` via the offline domain check.